requirements.txt
# Core dependencies
pypdf>=4.0.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.9.0
sentence-transformers>=3.0.0

# LLM Integration (LangChain)
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-huggingface>=1.0.0
langchain-ollama>=1.0.0

# UI
streamlit>=1.40.0

# Utilities
python-dotenv>=1.0.0
numpy>=2.0.0
tqdm>=4.67.3
